backend_auth/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
python-dotenv
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.5.0
langchain-huggingface>=0.1.0
langchain-openai>=0.1.20
chromadb>=0.4.0
pypdf>=3.0.0
python-docx>=1.0.0
pydantic-settings>=2.0.0
aiofiles>=23.0.0
